requirements.txt
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==4.3.0
blinker==1.7.0
certifi==2025.1.31
chardet==5.2.0
click==8.1.6
colorama==0.4.6
configobj==5.0.8
constantly==23.10.4
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
django-extensions==4.1
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
idna==3.10
pycparser==2.22
PyJWT==2.9.0
Pygments==2.19.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0